Move component ActiveTokyoAlert from 'components' to 'components/index/SiteTopUpper/TokyoAlert'
