+++ components/index/SiteTopUpper/TokyoAlertCard.vue
@@ -20,7 +20,7 @@
 <script lang="ts">
 import Vue from 'vue'
 
-import ActiveTokyoAlert from '@/components/ActiveTokyoAlert.vue'
+import ActiveTokyoAlert from '@/components/index/SiteTopUpper/TokyoAlert/ActiveTokyoAlert.vue'
 import tokyoAlert from '@/data/tokyo_alert.json'
 
 export default Vue.extend({
